+++ crates/objc2/src/__framework_prelude.rs
@@ -12,7 +12,7 @@ pub use std::os::raw::{
 };
 
 pub use crate::encode::{Encode, Encoding, RefEncode};
-pub use crate::ffi::{NSInteger, NSIntegerMax, NSUInteger, NSUIntegerMax, IMP};
+pub use crate::ffi::{NSInteger, NSIntegerMax, NSUInteger, NSUIntegerMax};
 pub use crate::mutability::{
     Immutable, ImmutableWithMutableSubclass, InteriorMutable, InteriorMutableWithSubclass,
     InteriorMutableWithSuperclass, IsIdCloneable, IsMainThreadOnly, IsRetainable, MainThreadOnly,
@@ -32,3 +32,4 @@ pub type AnyProtocol = AnyObject;
 pub type TodoFunction = *const c_void;
 pub type TodoClass = AnyObject;
 pub type TodoProtocols = AnyObject;
+pub type IMP = Option<crate::runtime::Imp>;