+++ crates/objc2/src/ffi/mod.rs
@@ -279,6 +279,9 @@ pub type objc_class = crate::runtime::AnyClass;
 #[deprecated = "merged with `runtime::AnyObject`"]
 pub type objc_object = crate::runtime::AnyObject;
 
+#[deprecated = "merged with `runtime::Imp`, and made non-null"]
+pub type IMP = Option<crate::runtime::Imp>;
+
 /// We don't know much about the actual structs, so better mark them `!Send`,
 /// `!Sync`, `!UnwindSafe`, `!RefUnwindSafe`, `!Unpin` and as mutable behind
 /// shared references.